:truck: [Rename] 메일발송 클래스의 패키지 분리 및 이동 #103
