+++ server/src/main/java/com/photoday/photoday/mail/config/EmailConfiguration.java
@@ -1,4 +1,4 @@
-package com.photoday.photoday.mail;
+package com.photoday.photoday.mail.config;
 
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
@@ -29,7 +29,6 @@ public class EmailConfiguration {
         props.put("mail.transport.protocol", "smtp");
         props.put("mail.smtp.auth", "true");
         props.put("mail.smtp.starttls.enable", "true");
-        props.put("mail.debug", "true");
 
         return mailSender;
     }
