+++ server/src/main/java/com/photoday/photoday/mail/config/SimpleEmailSendable.java
@@ -1,11 +1,15 @@
-package com.photoday.photoday.mail;
+package com.photoday.photoday.mail.config;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.mail.SimpleMailMessage;
 import org.springframework.mail.javamail.JavaMailSender;
 import org.springframework.stereotype.Component;
+
+import static io.netty.util.ResourceLeakDetector.setLevel;
 
 @Slf4j
 @Component
